Speciation of emissions

speciate separates emissions in different compounds. It covers black carbon and organic matter from particulate matter. Soon it will be added more speciations

spec_industry(x, spec = "GAS", code, usepa_code, list = FALSE)

Arguments

x
Emissions estimation

spec
Character any of "GAS" or "PM"

code
Character any of "01_petroleum", "02_portland_cement", "03_paper_cellulose", "04_bagasse", "05_natural_gas", "06_coal" OR "07_fuel_oil"

usepa_code
Character any of "4713", "3011", "8799", "SUG02", "2442", "1185", "0001", "13505C", "91127", "4733", "SUGP02", "1120310" OR "91114"

list
when TRUE returns a list with number of elements of the list as the number species of pollutants

Value

dataframe or list

References

Simon, H., Beck, L., Bhave, P. V., Divita, F., Hsu, Y., Luecken, D., ... Strum, M. (2010). The development and uses of EPA SPECIATE database. Atmospheric Pollution Research, 1(4), 196-206.

Examples

{
# Do not run
pm <- rnorm(n = 100, mean = 400, sd = 2)
(df <- spec_industry(pm, code = "01_petroleum"))
}

#>                x                                   species
#>     1: 5.7527964 1-(1,1-dimethylethyl)-3,5-dimethylbenzene
#>     2: 5.8007468 1-(1,1-dimethylethyl)-3,5-dimethylbenzene
#>     3: 5.7227515 1-(1,1-dimethylethyl)-3,5-dimethylbenzene
#>     4: 5.7931897 1-(1,1-dimethylethyl)-3,5-dimethylbenzene
#>     5: 5.8113555 1-(1,1-dimethylethyl)-3,5-dimethylbenzene
#>    ---                                                    
#> 14496: 0.2086360                            vinyl chloride
#> 14497: 0.2065388                            vinyl chloride
#> 14498: 0.2078844                            vinyl chloride
#> 14499: 0.2082677                            vinyl chloride
#> 14500: 0.2065985                            vinyl chloride